move useRemainingNativeTxs to own file
